Fancy up the colorize button
