+++ frontend/src/screens/App/screens/ViewerPane/components/ColorLayer.less
@@ -1,3 +1,5 @@
+@import '../../../../../shared/styles/colors.less';
+
 :local {
   .unfaded {
     opacity: 1;
