+++ frontend/src/screens/App/screens/ViewerPane/components/ColorizeButton.less
@@ -1,0 +1,43 @@
+@import '../../../../../shared/styles/colors.less';
+
+@enabled-color: @blue;
+
+.colorizeButton {
+  animation: loading 10s linear infinite paused;
+
+  @keyframes loading {
+    0% {
+      background-position-x: 600%;
+    }
+    100% {
+      background-position-x: 0%;
+    }
+  }
+}
+
+.enabled {
+  @red-softer: #b36254;
+  border-color: @enabled-color !important;
+
+  background: @enabled-color
+    linear-gradient(
+      to right,
+      @blue ((0/3) * 100%),
+      @red-softer ((1/3) * 100%),
+      @green ((2/3) * 100%),
+      @blue ((3/3) * 100%)
+    )
+    100% 0% / 600% 100%;
+
+  &.loading {
+    animation-play-state: running;
+  }
+}
+
+.magicIcon {
+  height: 1.9ex;
+  filter: invert(1);
+  margin-right: 0.5em;
+  margin-bottom: -0.3ex;
+  margin-left: -0.15em;
+}
