+++ frontend/src/screens/App/screens/ViewerPane/components/ImageButtons.tsx
@@ -5,16 +5,11 @@ import { API_BASE } from 'utils/apiConstants';
 
 import Button, { ButtonStyledLink } from 'shared/components/Button';
 import { useStoryDraftStore } from '../../SubmitStoryWizard';
-import { useFeatureFlag } from 'screens/App/shared/stores/FeatureFlagsStore';
-import FeatureFlag from 'screens/App/shared/types/FeatureFlag';
-import useColorizationStore from '../stores/ColorizationStore';
+import ColorizeButton from './ColorizeButton';
 
 export default function ImageButtons(): JSX.Element {
   const { identifier: photoIdentifier } = useParams<{ identifier?: string }>();
   const initializeStoryDraft = useStoryDraftStore((state) => state.initialize);
-  const colorizationEnabled = useFeatureFlag(FeatureFlag.COLORIZATION);
-  const { toggleColorization, isLoading, colorEnabledForIdentifier } =
-    useColorizationStore();
 
   return (
     <div>
@@ -27,19 +22,7 @@ export default function ImageButtons(): JSX.Element {
       >
         Order Print
       </ButtonStyledLink>
-      {colorizationEnabled && (
-        <Button
-          buttonTheme="viewer"
-          buttonStyle="secondary"
-          disabled={isLoading}
-          onClick={() => {
-            toggleColorization(photoIdentifier);
-          }}
-        >
-          Color{isLoading && '...'} (
-          {photoIdentifier === colorEnabledForIdentifier ? 'On' : 'Off'})
-        </Button>
-      )}
+      <ColorizeButton photoIdentifier={photoIdentifier} />
       <Button
         buttonTheme="viewer"
         buttonStyle="secondary"
